+++ skate-plugin/src/main/kotlin/com/slack/sgp/intellij/codeowners/model/CodeOwner.kt
@@ -17,4 +17,8 @@ package com.slack.sgp.intellij.codeowners.model
 
 import kotlinx.serialization.Serializable
 
-@Serializable data class CodeOwner(val name: String, val paths: List<String>)
+@Serializable
+data class CodeOwner(
+  val name: String,
+  val paths: List<@Serializable(with = PathObjectSerializer::class) Path>,
+)