+++ slack-plugin/src/main/kotlin/slack/gradle/SlackRootPlugin.kt
@@ -146,7 +146,7 @@ internal class SlackRootPlugin @Inject constructor(private val buildFeatures: Bu
     if (slackProperties.strictJdk) {
       val runtimeVersion =
         project.providers.systemProperty("java.specification.version").get().toInt()
-      val jdk = slackProperties.jdkVersion
+      val jdk = slackProperties.versions.jdk.get()
       check(jdk == runtimeVersion) {
         """
           Current Java version ($runtimeVersion) does not match the enforced version ($jdk).
@@ -168,7 +168,8 @@ internal class SlackRootPlugin @Inject constructor(private val buildFeatures: Bu
       project.configureGit(slackProperties)
     }
     project.configureSlackRootBuildscript(
-      slackProperties.jvmVendor.map(JvmVendorSpec::matching).orNull
+      slackProperties.versions.jdk.asProvider(project.providers),
+      slackProperties.jvmVendor.map(JvmVendorSpec::matching).orNull,
     )
     LintTasks.configureRootProject(project)
     DetektTasks.configureRootProject(project, slackProperties)
@@ -467,10 +468,13 @@ internal class SlackRootPlugin @Inject constructor(private val buildFeatures: Bu
   }
 }
 
-private fun Project.configureSlackRootBuildscript(jvmVendor: JvmVendorSpec?) {
+private fun Project.configureSlackRootBuildscript(
+  jdkProvider: Provider<Int>,
+  jvmVendor: JvmVendorSpec?,
+) {
   // Only register bootstrap if explicitly requested for now
   if (CoreBootstrapTask.isBootstrapEnabled(this)) {
-    CoreBootstrapTask.register(this, jvmVendor)
+    CoreBootstrapTask.register(this, jdkProvider, jvmVendor)
   }
   InstallCommitHooksTask.register(this)
 }