+++ slack-plugin/src/test/kotlin/slack/fakes/GradleStub.kt
@@ -27,6 +27,7 @@ import org.gradle.api.execution.TaskExecutionGraph
 import org.gradle.api.initialization.IncludedBuild
 import org.gradle.api.initialization.Settings
 import org.gradle.api.invocation.Gradle
+import org.gradle.api.invocation.GradleLifecycle
 import org.gradle.api.plugins.ExtensionContainer
 import org.gradle.api.plugins.ObjectConfigurationAction
 import org.gradle.api.plugins.PluginContainer
@@ -77,6 +78,10 @@ class GradleStub(private val startParameter: StartParameter) : Gradle {
   }
 
   override fun removeProjectEvaluationListener(listener: ProjectEvaluationListener) = Unit
+
+  override fun getLifecycle(): GradleLifecycle {
+    TODO("Not yet implemented")
+  }
 
   override fun beforeSettings(closure: Closure<*>) = Unit